Recipe Microservice

+++ sts-workspace/foodie-app/src/main/java/com/del/foodieapp/service/MenuItemServiceImpl.java
@@ -5,13 +5,20 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.client.RestTemplate;
 
+import com.del.foodieapp.exception.ItemNotFoundException;
 import com.del.foodieapp.model.ItemCategory;
 import com.del.foodieapp.model.MenuItem;
+import com.del.foodieapp.model.Recipe;
 import com.del.foodieapp.repository.MenuItemRepository;
 
 @Service
 public class MenuItemServiceImpl implements MenuItemService {
+	
+	private String RECIPE_SERVICE_URL = "http://localhost:8200/recipe/";
+	
+	RestTemplate recipeService = new RestTemplate();
 	
 	@Autowired
 	private MenuItemRepository repo;
@@ -23,8 +30,13 @@ public class MenuItemServiceImpl implements MenuItemService {
 
 	@Override
 	public MenuItem getItemByCode(long itemCode) {
-		
-		return repo.findById(itemCode).get();
+		if(!repo.existsById(itemCode)) {
+			throw new ItemNotFoundException("Menu Item with code "+itemCode+" not found");
+		}
+		MenuItem item =  repo.findById(itemCode).get();
+		Recipe recipe = getRecipce(item.getItemName());
+		item.setRecipe(recipe);
+		return item;
 	}
 
 	@Override
@@ -39,7 +51,9 @@ public class MenuItemServiceImpl implements MenuItemService {
 
 	@Transactional
 	public void deleteItem(long itemCode) {
-		
+		if(!repo.existsById(itemCode)) {
+			throw new ItemNotFoundException("Menu Item with code "+itemCode+" not found");
+		}
 		repo.deleteById(itemCode);
 	}
 	
@@ -53,6 +67,10 @@ public class MenuItemServiceImpl implements MenuItemService {
 		ItemCategory c = ItemCategory.valueOf(category);
 		
 		return repo.findByCategory(c);
+	}
+	
+	public Recipe getRecipce(String itemName) {
+		return recipeService.getForObject(RECIPE_SERVICE_URL+itemName, Recipe.class);
 	}
 
 }
